build(log): Make log initialization optional
