+++ relay-config/Cargo.toml
@@ -19,7 +19,7 @@ human-size = "0.4.1"
 num_cpus = "1.13.0"
 relay-auth = { path = "../relay-auth" }
 relay-common = { path = "../relay-common" }
-relay-log = { path = "../relay-log" }
+relay-log = { path = "../relay-log", features = ["init"] }
 relay-redis = { path = "../relay-redis" }
 serde = { version = "1.0.114", features = ["derive"] }
 serde_json = "1.0.55"
