+++ relay-log/Cargo.toml
@@ -10,12 +10,26 @@ license-file = "../LICENSE"
 publish = false
 
 [dependencies]
-chrono = "0.4.19"
-console = "0.10.0"
-env_logger = "0.7.1"
+chrono = { version = "0.4.19", optional = true }
+console = { version = "0.10.0", optional = true }
+env_logger = { version = "0.7.1", optional = true }
 failure = "0.1.8"
 log = { version = "0.4.11", features = ["serde"] }
-pretty_env_logger = "0.4.0"
-sentry = { version = "0.22.0", features = ["debug-images", "log"] }
-serde = { version = "1.0.114", features = ["derive"] }
-serde_json = "1.0.55"
+pretty_env_logger = { version = "0.4.0", optional = true }
+sentry = { version = "0.22.0", features = ["debug-images", "log"], optional = true }
+sentry-core = { version = "0.22.0" }
+serde = { version = "1.0.114", features = ["derive"], optional = true }
+serde_json = { version = "1.0.55", optional = true }
+
+[features]
+default = []
+test = ["env_logger"]
+init = [
+    "chrono",
+    "console",
+    "env_logger",
+    "pretty_env_logger",
+    "sentry",
+    "serde",
+    "serde_json"
+]
