+++ relay/Cargo.toml
@@ -26,7 +26,7 @@ hostname = "0.3.1"
 lazy_static = "1.4.0"
 relay-common = { path = "../relay-common" }
 relay-config = { path = "../relay-config" }
-relay-log = { path = "../relay-log" }
+relay-log = { path = "../relay-log", features = ["init"] }
 relay-server = { path = "../relay-server" }
 
 [build-dependencies]
